Clarify --dir requirements in README.md
Since Ventura is pretty new, I think it'd be good to warn people before they download a large image. `tart run --help` gives me this:

```
  --dir <name:path[:ro]>  Additional directory shares with an optional read-only specifier
                          (e.g. --dir="build:~/src/build" --dir="sources:~/src/sources:ro")
        Requires host to be macOS 13.0 (Ventura) or newer.
        All shared directories are automatically mounted to "/Volumes/My Shared Files" directory on macOS,
        while on Linux you have to do it manually: "mount -t virtiofs com.apple.virtio-fs.automount /mount/point".
        For macOS guests, they must be running macOS 13.0 (Ventura) or newer.
```

I'm unclear if `--dir` requires both host and guest to be running Ventura or if the "For macOS guests..." is only talking about the "All shared directories..." portion.