Use specified pixels when refit is disabled.
When display refit has been explicitly disabled by a user, we should
prefer using the pixel-based constructor to ensure that the resolution
does not get changed inadvertantly.

A common scenario is running the VMs in headless mode without GUI,
however even in those cases the screen APIs will return a value for the
main screen so we can't solely rely on that as the test criteria for
opting out of the relatively scaled display configuration.